Adding HTTP Header generator that results in more code coverage. Also harness for fuzzing HTTP11Processor.service
The harness fails because I do not load a webapp into Tomcat. It gets through all the processing code and generation of a Request/Response object.
Before these objects are passed to the webapp, we can mutate amd simply re-call the service function to see if we get the same mutated object.
We will have to insert tomcat-specific instrumentation for this.