Working JQF instrumentation for Simple Buggy Servlet
Ties together the Custom Generators and Tomcat to fuzz a simple, stupid webapp.
It is able to spin up a Tomcat instance, attach to the execution thread for
the webapp, and get coverage within the webapp itself, it is even able to find
a crash within the webapp within a few seconds. The Zest fuzzer runs at an
execution speed of about 14-15 execs/second.